Adjusted the powers, and commited a few things I forgot to.  hehe... richard would kill me if he didn't have 'em.
git-svn-id: svn://10.0.50.55/puffbomb/trunk@1823 d14347b8-48fd-0310-9e84-97907d434d85

+++ Code/Game/PuffBOMB/Objects/BasicBomb/BasicBomb.cpp
@@ -43,9 +43,9 @@ void cBasicBomb::Work() {
 			Engine2D::cImpulse(
 				Pos,
 				// Inner Radius, Intensity, Tangent //
-				Real( 0 ), Real( 25 ), Real( 0 ),
+				Real( 0 ), Real( 50 ), Real( 0 ),
 				// Outer Radius, Intensity, Tangent //
-				Real( 1512 ), Real( 0 ), Real( 0 )
+				Real( 512 ), Real( 0 ), Real( 0 )
 				)
 			);
 	}
